Refactor eosio_system_tester into header file. GH #3039
