+++ unittests/eosio.system_tests.cpp
@@ -1,422 +1,14 @@
 #include <boost/test/unit_test.hpp>
-#include <eosio/testing/tester.hpp>
-#include <eosio/chain/abi_serializer.hpp>
 #include <eosio/chain/contract_table_objects.hpp>
 #include <eosio/chain/global_property_object.hpp>
 #include <eosio/chain/resource_limits.hpp>
 #include <eosio/chain/wast_to_wasm.hpp>
 
-#include <eosio.system/eosio.system.wast.hpp>
-#include <eosio.system/eosio.system.abi.hpp>
-
-#include <eosio.token/eosio.token.wast.hpp>
-#include <eosio.token/eosio.token.abi.hpp>
-
-#include <eosio.msig/eosio.msig.wast.hpp>
-#include <eosio.msig/eosio.msig.abi.hpp>
-
 #include <Runtime/Runtime.h>
 
-#include <fc/variant_object.hpp>
+#include "eosio_system_tester.hpp"
 
-#ifdef NON_VALIDATING_TEST
-#define TESTER tester
-#else
-#define TESTER validating_tester
-#endif
-
-using namespace eosio::testing;
-using namespace eosio;
-using namespace eosio::chain;
-using namespace eosio::testing;
-using namespace fc;
-
-using mvo = fc::mutable_variant_object;
-
-class eosio_system_tester : public TESTER {
-public:
-
-   eosio_system_tester() {
-      produce_blocks( 2 );
-
-      create_accounts( { N(eosio.token) } );
-
-      produce_blocks( 100 );
-
-      set_code( N(eosio.token), eosio_token_wast );
-      set_abi( N(eosio.token), eosio_token_abi );
-
-      {
-         const auto& accnt = control->db().get<account_object,by_name>( N(eosio.token) );
-         abi_def abi;
-         BOOST_REQUIRE_EQUAL(abi_serializer::to_abi(accnt.abi, abi), true);
-         token_abi_ser.set_abi(abi);
-      }
-
-      create_currency( N(eosio.token), config::system_account_name, core_from_string("10000000000.0000") );
-      issue(config::system_account_name,      core_from_string("1000000000.0000"));
-      BOOST_REQUIRE_EQUAL( core_from_string("1000000000.0000"), get_balance( "eosio" ) );
-
-      set_code( config::system_account_name, eosio_system_wast );
-      set_abi( config::system_account_name, eosio_system_abi );
-
-      {
-         const auto& accnt = control->db().get<account_object,by_name>( config::system_account_name );
-         abi_def abi;
-         BOOST_REQUIRE_EQUAL(abi_serializer::to_abi(accnt.abi, abi), true);
-         abi_ser.set_abi(abi);
-      }
-
-      produce_blocks();
-
-      create_account_with_resources( N(alice1111111), N(eosio), core_from_string("1.0000"), false );
-      create_account_with_resources( N(bob111111111), N(eosio), core_from_string("0.4500"), false );
-      create_account_with_resources( N(carol1111111), N(eosio), core_from_string("1.0000"), false );
-
-
-      BOOST_REQUIRE_EQUAL( core_from_string("1000000000.0000"), get_balance( "eosio" ) );
-   }
-
-
-   void create_accounts_with_resources( vector<account_name> accounts, account_name creator = N(eosio) ) {
-      for( auto a : accounts ) {
-         create_account_with_resources( a, creator );
-      }
-   }
-
-   transaction_trace_ptr create_account_with_resources( account_name a, account_name creator ) {
-      signed_transaction trx;
-      set_transaction_headers(trx);
-
-      authority owner_auth;
-      owner_auth =  authority( get_public_key( a, "owner" ) );
-
-      trx.actions.emplace_back( vector<permission_level>{{creator,config::active_name}},
-                                newaccount{
-                                   .creator  = creator,
-                                   .name     = a,
-                                   .owner    = owner_auth,
-                                   .active   = authority( get_public_key( a, "active" ) )
-                                });
-
-      trx.actions.emplace_back( get_action( N(eosio), N(buyrambytes), vector<permission_level>{{creator,config::active_name}},
-                                            mvo()
-                                            ("payer", creator)
-                                            ("receiver", a)
-                                            ("bytes", 8000) )
-                              );
-      trx.actions.emplace_back( get_action( N(eosio), N(delegatebw), vector<permission_level>{{creator,config::active_name}},
-                                            mvo()
-                                            ("from", creator)
-                                            ("receiver", a)
-                                            ("stake_net_quantity", core_from_string("10.0000") )
-                                            ("stake_cpu_quantity", core_from_string("10.0000") )
-                                            ("transfer", 0 )
-                                          )
-                                );
-
-      set_transaction_headers(trx);
-      trx.sign( get_private_key( creator, "active" ), chain_id_type()  );
-      return push_transaction( trx );
-   }
-
-   transaction_trace_ptr create_account_with_resources( account_name a, account_name creator, asset ramfunds, bool multisig,
-                                                        asset net = core_from_string("10.0000"), asset cpu = core_from_string("10.0000") ) {
-      signed_transaction trx;
-      set_transaction_headers(trx);
-
-      authority owner_auth;
-      if (multisig) {
-         // multisig between account's owner key and creators active permission
-         owner_auth = authority(2, {key_weight{get_public_key( a, "owner" ), 1}}, {permission_level_weight{{creator, config::active_name}, 1}});
-      } else {
-         owner_auth =  authority( get_public_key( a, "owner" ) );
-      }
-
-      trx.actions.emplace_back( vector<permission_level>{{creator,config::active_name}},
-                                newaccount{
-                                   .creator  = creator,
-                                   .name     = a,
-                                   .owner    = owner_auth,
-                                   .active   = authority( get_public_key( a, "active" ) )
-                                });
-
-      trx.actions.emplace_back( get_action( N(eosio), N(buyram), vector<permission_level>{{creator,config::active_name}},
-                                            mvo()
-                                            ("payer", creator)
-                                            ("receiver", a)
-                                            ("quant", ramfunds) )
-                              );
-
-      trx.actions.emplace_back( get_action( N(eosio), N(delegatebw), vector<permission_level>{{creator,config::active_name}},
-                                            mvo()
-                                            ("from", creator)
-                                            ("receiver", a)
-                                            ("stake_net_quantity", net )
-                                            ("stake_cpu_quantity", cpu )
-                                            ("transfer", 0 )
-                                          )
-                                );
-
-      set_transaction_headers(trx);
-      trx.sign( get_private_key( creator, "active" ), chain_id_type()  );
-      return push_transaction( trx );
-   }
-
-   transaction_trace_ptr setup_producer_accounts( const std::vector<account_name>& accounts ) {
-      account_name creator(N(eosio));
-      signed_transaction trx;
-      set_transaction_headers(trx);
-      asset cpu = core_from_string("80.0000");
-      asset net = core_from_string("80.0000");
-      asset ram = core_from_string("1.0000");
-
-      for (const auto& a: accounts) {
-         authority owner_auth( get_public_key( a, "owner" ) );
-         trx.actions.emplace_back( vector<permission_level>{{creator,config::active_name}},
-                                   newaccount{
-                                         .creator  = creator,
-                                         .name     = a,
-                                         .owner    = owner_auth,
-                                         .active   = authority( get_public_key( a, "active" ) )
-                                         });
-
-         trx.actions.emplace_back( get_action( N(eosio), N(buyram), vector<permission_level>{ {creator, config::active_name} },
-                                               mvo()
-                                               ("payer", creator)
-                                               ("receiver", a)
-                                               ("quant", ram) )
-                                   );
-
-         trx.actions.emplace_back( get_action( N(eosio), N(delegatebw), vector<permission_level>{ {creator, config::active_name} },
-                                               mvo()
-                                               ("from", creator)
-                                               ("receiver", a)
-                                               ("stake_net_quantity", net)
-                                               ("stake_cpu_quantity", cpu )
-                                               ("transfer", 0 )
-                                               )
-                                   );
-      }
-
-      set_transaction_headers(trx);
-      trx.sign( get_private_key( creator, "active" ), chain_id_type()  );
-      return push_transaction( trx );
-   }
-
-   action_result buyram( const account_name& payer, account_name receiver, const asset& eosin ) {
-      return push_action( payer, N(buyram), mvo()( "payer",payer)("receiver",receiver)("quant",eosin) );
-   }
-   action_result buyrambytes( const account_name& payer, account_name receiver, uint32_t numbytes ) {
-      return push_action( payer, N(buyrambytes), mvo()( "payer",payer)("receiver",receiver)("bytes",numbytes) );
-   }
-
-   action_result sellram( const account_name& account, uint64_t numbytes ) {
-      return push_action( account, N(sellram), mvo()( "account", account)("bytes",numbytes) );
-   }
-
-   action_result push_action( const account_name& signer, const action_name &name, const variant_object &data, bool auth = true ) {
-         string action_type_name = abi_ser.get_action_type(name);
-
-         action act;
-         act.account = config::system_account_name;
-         act.name = name;
-         act.data = abi_ser.variant_to_binary( action_type_name, data );
-
-         return base_tester::push_action( std::move(act), auth ? uint64_t(signer) : signer == N(bob111111111) ? N(alice1111111) : N(bob111111111) );
-   }
-
-   action_result stake( const account_name& from, const account_name& to, const asset& net, const asset& cpu ) {
-      return push_action( name(from), N(delegatebw), mvo()
-                          ("from",     from)
-                          ("receiver", to)
-                          ("stake_net_quantity", net)
-                          ("stake_cpu_quantity", cpu)
-                          ("transfer", 0 )
-      );
-   }
-
-   action_result stake( const account_name& acnt, const asset& net, const asset& cpu ) {
-      return stake( acnt, acnt, net, cpu );
-   }
-
-   action_result stake_with_transfer( const account_name& from, const account_name& to, const asset& net, const asset& cpu ) {
-      return push_action( name(from), N(delegatebw), mvo()
-                          ("from",     from)
-                          ("receiver", to)
-                          ("stake_net_quantity", net)
-                          ("stake_cpu_quantity", cpu)
-                          ("transfer", true )
-      );
-   }
-
-   action_result stake_with_transfer( const account_name& acnt, const asset& net, const asset& cpu ) {
-      return stake_with_transfer( acnt, acnt, net, cpu );
-   }
-
-   action_result unstake( const account_name& from, const account_name& to, const asset& net, const asset& cpu ) {
-      return push_action( name(from), N(undelegatebw), mvo()
-                          ("from",     from)
-                          ("receiver", to)
-                          ("unstake_net_quantity", net)
-                          ("unstake_cpu_quantity", cpu)
-      );
-   }
-
-   action_result unstake( const account_name& acnt, const asset& net, const asset& cpu ) {
-      return unstake( acnt, acnt, net, cpu );
-   }
-
-   static fc::variant_object producer_parameters_example( int n ) {
-      return mutable_variant_object()
-         ("max_block_net_usage", 10000000 + n )
-         ("target_block_net_usage_pct", 10 + n )
-         ("max_transaction_net_usage", 1000000 + n )
-         ("base_per_transaction_net_usage", 100 + n)
-         ("net_usage_leeway", 500 + n )
-         ("context_free_discount_net_usage_num", 1 + n )
-         ("context_free_discount_net_usage_den", 100 + n )
-         ("max_block_cpu_usage", 10000000 + n )
-         ("target_block_cpu_usage_pct", 10 + n )
-         ("max_transaction_cpu_usage", 1000000 + n )
-         ("min_transaction_cpu_usage", 100 + n )
-         ("max_transaction_lifetime", 3600 + n)
-         ("deferred_trx_expiration_window", 600 + n)
-         ("max_transaction_delay", 10*86400+n)
-         ("max_inline_action_size", 4096 + n)
-         ("max_inline_action_depth", 4 + n)
-         ("max_authority_depth", 6 + n)
-         ("max_generated_transaction_count", 10 + n)
-         ("max_ram_size", (n % 10 + 1) * 1024 * 1024)
-         ("ram_reserve_ratio", 100 + n);
-   }
-
-   action_result regproducer( const account_name& acnt, int params_fixture = 1 ) {
-      action_result r = push_action( acnt, N(regproducer), mvo()
-                          ("producer",  acnt )
-                          ("producer_key", get_public_key( acnt, "active" ) )
-                          ("url", "" )
-                          ("location", 0 )
-      );
-      BOOST_REQUIRE_EQUAL( success(), r);
-      return r;
-   }
-
-   uint32_t last_block_time() const {
-      return time_point_sec( control->head_block_time() ).sec_since_epoch();
-   }
-
-   asset get_balance( const account_name& act ) {
-
-      vector<char> data = get_row_by_account( N(eosio.token), act, N(accounts), symbol(CORE_SYMBOL).to_symbol_code().value );
-      return data.empty() ? asset(0, symbol(CORE_SYMBOL)) : token_abi_ser.binary_to_variant("account", data)["balance"].as<asset>();
-   }
-
-   fc::variant get_total_stake( const account_name& act ) {
-      vector<char> data = get_row_by_account( config::system_account_name, act, N(userres), act );
-      return data.empty() ? fc::variant() : abi_ser.binary_to_variant( "user_resources", data );
-   }
-
-   fc::variant get_voter_info( const account_name& act ) {
-      vector<char> data = get_row_by_account( config::system_account_name, config::system_account_name, N(voters), act );
-      return data.empty() ? fc::variant() : abi_ser.binary_to_variant( "voter_info", data );
-   }
-
-   fc::variant get_producer_info( const account_name& act ) {
-      vector<char> data = get_row_by_account( config::system_account_name, config::system_account_name, N(producers), act );
-      return abi_ser.binary_to_variant( "producer_info", data );
-   }
-
-   void create_currency( name contract, name manager, asset maxsupply ) {
-      auto act =  mutable_variant_object()
-         ("issuer",       manager )
-         ("maximum_supply", maxsupply );
-
-      base_tester::push_action(contract, N(create), contract, act );
-   }
-
-   void issue( name to, const asset& amount, name manager = config::system_account_name ) {
-      base_tester::push_action( N(eosio.token), N(issue), manager, mutable_variant_object()
-                                ("to",      to )
-                                ("quantity", amount )
-                                ("memo", "")
-                                );
-   }
-   void transfer( name from, name to, const asset& amount, name manager = config::system_account_name ) {
-      base_tester::push_action( N(eosio.token), N(transfer), manager, mutable_variant_object()
-                                ("from",    from)
-                                ("to",      to )
-                                ("quantity", amount)
-                                ("memo", "")
-                                );
-   }
-
-   double stake2votes( asset stake ) {
-      auto now = control->pending_block_time().time_since_epoch().count() / 1000000;
-      return stake.amount * pow(2, int64_t((now - (config::block_timestamp_epoch / 1000)) / (86400 * 7))/ double(52) ); // 52 week periods (i.e. ~years)
-   }
-
-   double stake2votes( const string& s ) {
-      return stake2votes( core_from_string(s) );
-   }
-
-   fc::variant get_stats( const string& symbolname ) {
-      auto symb = eosio::chain::symbol::from_string(symbolname);
-      auto symbol_code = symb.to_symbol_code().value;
-      vector<char> data = get_row_by_account( N(eosio.token), symbol_code, N(stat), symbol_code );
-      return data.empty() ? fc::variant() : token_abi_ser.binary_to_variant( "currency_stats", data );
-   }
-
-   asset get_token_supply() {
-      return get_stats("4," CORE_SYMBOL_NAME)["supply"].as<asset>();
-   }
-
-   fc::variant get_global_state() {
-      vector<char> data = get_row_by_account( N(eosio), N(eosio), N(global), N(global) );
-      if (data.empty()) std::cout << "\nData is empty\n" << std::endl;
-      return data.empty() ? fc::variant() : abi_ser.binary_to_variant( "eosio_global_state", data );
-
-   }
-
-   fc::variant get_refund_request( name account ) {
-      vector<char> data = get_row_by_account( N(eosio), account, N(refunds), account );
-      return data.empty() ? fc::variant() : abi_ser.binary_to_variant( "refund_request", data );
-   }
-
-   abi_serializer abi_ser;
-   abi_serializer token_abi_ser;
-};
-
-fc::mutable_variant_object voter( account_name acct ) {
-   return mutable_variant_object()
-      ("owner", acct)
-      ("proxy", name(0).to_string())
-      ("producers", variants() )
-      ("staked", int64_t(0))
-      //("last_vote_weight", double(0))
-      ("proxied_vote_weight", double(0))
-      ("is_proxy", 0)
-      ("deferred_trx_id", 0)
-      ("last_unstake_time", fc::time_point_sec() )
-      ("unstaking", asset() )
-      ;
-}
-
-fc::mutable_variant_object voter( account_name acct, const asset& vote_stake ) {
-   return voter( acct )( "staked", vote_stake.amount );
-}
-
-fc::mutable_variant_object voter( account_name acct, int64_t vote_stake ) {
-   return voter( acct )( "staked", vote_stake );
-}
-
-fc::mutable_variant_object proxy( account_name acct ) {
-   return voter( acct )( "is_proxy", 1 );
-}
-
-inline uint64_t M( const string& eos_str ) {
-   return core_from_string( eos_str ).amount;
-}
+using namespace eosio_system;
 
 BOOST_AUTO_TEST_SUITE(eosio_system_tests)
